+++ internal/pkg/printer/printer.go
@@ -1,12 +1,27 @@
 package printer
 
 import (
+	"fmt"
 	"github.com/mitchellh/colorstring"
 	"io"
 	"os"
 )
 
 var writer io.Writer = os.Stdout
+
+var coloriser colorstring.Colorize
+
+func init() {
+	coloriser = colorstring.Colorize{
+		Colors:  colorstring.DefaultColors,
+		Disable: false,
+		Reset:   true,
+	}
+}
+
+func Disable() {
+	coloriser.Disable = true
+}
 
 func SetOutput(out io.Writer) {
 	writer = out
@@ -15,13 +30,13 @@ func SetOutput(out io.Writer) {
 // Valid colour codes are listed at: https://github.com/mitchellh/colorstring/blob/master/colorstring.go
 
 func Fprint(text string) (int, error) {
-	return colorstring.Fprint(writer, text)
+	return fmt.Fprint(writer, coloriser.Color(text))
 }
 
 func Fprintf(format string, args ...interface{}) (int, error) {
-	return colorstring.Fprintf(writer, format, args...)
+	return fmt.Fprintf(writer, coloriser.Color(format), args...)
 }
 
 func Fprintln(text string) (int, error) {
-	return colorstring.Fprintln(writer, text)
+	return fmt.Fprintln(writer, coloriser.Color(text))
 }